+++ context/DocPage-context.js
@@ -21,12 +21,15 @@ function CurrentDocProvider({ children }) {
     console.log('total Doc',docState);
   });
   
+
+  // Updates the state fo the docs
   function UpdateDocState(event) {
     if (!event) {
       return new Error('Require a name and value attribute to update the Doc state')
     }
-    const name = event.name;
-    const value = event.value;
+    const {name,value} = event;
+
+
     setDocState({ ...docState, [name]: value })
     console.log(docState)
    }
